Moved Uncompress to own package for easier import
